+++ metro_mini_firmware/lcd.h
@@ -14,9 +14,9 @@ class LCD : private LiquidCrystal_I2C
     void startup_screen(); // top row for lcd startup message
     void gpslock_screen(int sats); // top row for gps search screen
     void zero_prompt_screen(); // top row for when no 0s are set
-    void standard_screen(int zero, int meas); // top row for during measurement taking
-    void print_measurement(int zero, int meas, float x, float y, float z); // print out the measurements taken
-    void print_zero(int zero, float x, float y); // print out the location of the zero
+    void standard_screen(uint8_t zero, uint8_t meas); // top row for during measurement taking
+    void print_measurement(uint8_t zero, uint8_t meas, float x, float y, float z); // print out the measurements taken
+    void print_zero(uint8_t zero, float x, float y); // print out the location of the zero
   private:
     int loopcount = 0;
     char loop = '\\';
